pyproject.toml
[tool.poetry]
name = "repl-nix-workspace"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
google-cloud-vision = "^3.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
